+++ src/main/java/ru/kata/spring/boot_security/demo/model/Role.java
@@ -10,7 +10,7 @@ import java.util.Objects;
 @Table(name = "role")
 public class Role implements GrantedAuthority {
     @Id
-    private int id;
+    private Long id;
 
     @Column
     private String name;
@@ -18,16 +18,16 @@ public class Role implements GrantedAuthority {
     public Role() {
     }
 
-    public Role(int id, String name) {
+    public Role(Long id, String name) {
         this.name = name;
         this.id = id;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
